feat: strategy navigation (#850)
* feat: strategy navigation

* fix: /strategies navigation

* fix: set tabId on load
