+++ packages/frontend/pages/strategies/bull.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import BullStrategy from '@components/Strategies/Bull'
+
+const Page: React.FC = () => <BullStrategy />
+
+export default Page
